test, validate items in bulk transacton client
